requirements.txt
# pip install -r requirements.txt

# Base ----------------------------------------
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=7.1.2
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0
Flask==2.0.2
Flask-Cors==3.0.10


# tensorflow -------------------------------------
tensorflow-gpu==2.1.0
keras==2.3.1
h5py==2.10.0

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0
